Remove tests of specific number of reports
Depending on test run order, there may be a different number of reports in tables and lists (like the "archived" list in the admin panel). Counting them is unreliable and not useful for testing.

Instead, tests should identify if the relevant report is present.

For example, rather than "There should be 4 complete reports listed", use "The report for client 12345678 should be in the list"