Upgrade to PHP 5.6, Alpine 3.8
Critically, we need to upgrade to Alpine 3.8 so that the apk version of `postgres-client` matches the version of database we use (9.6). The only Alpine images running PHP 5 and Alpine 3.8 use PHP 5.6, so we've had to upgrade there too.

Fortunately, this is only a minor upgrade and there are should be no breaking changes or deprecations that affect us.