Don't try and delete access token if infeasible
Somewhere between Symfony 3.4.11 and 3.4.28, there was a change that meant the order of logout listener being called may be different. This means that the user token may be an `AnonymousToken` by the time the logout listener is called. Therefore, we should check that the token is a `UsernamePasswordToken` before trying to invalidate it against the API.

We don't _need_ to manually invalidate the login token at logout because they automatically expire after an hour (as would happen if the user closed the tab rather than logging out).