Allow govuk-elements and govuk-frontend work side-by-side
We're not going to be able to make a straight swap, so need to ensure they can both be used at once.

This overwrites a rule in govuk-elements which was setting the `html` font-size to small. This conflicted with govuk-frontend's use of rems. The govuk-elements rule was just for IE6/7 compatibility, which is no longer an issue.